Fix crash with current version of Barotrauma
The security Slipsuit includes empty fabrication requirements.
This is probably due to it being a modified version of the combat
diving suit and would suggest some kind of recipe inheritance for
variant items?

Though I haven't tested this aspect for other variant items. Their
recipes may not be fully correct.